Rename subcommand modules removing the 'cli_' prefix
