+++ logilica_cli/__main__.py
@@ -8,8 +8,8 @@ from jsonschema import ValidationError
 import yaml
 
 from logilica_cli import sort_click_command_parameters
-from logilica_cli.cli_data_sources import data_sources
-from logilica_cli.cli_weekly_report import weekly_report
+from logilica_cli.data_sources import data_sources
+from logilica_cli.weekly_report import weekly_report
 from logilica_cli.configuration_schema import validate_configuration
 
 # Default values for command options
