+++ logilica_cli/weekly_report.py
